requirements.txt
numpy==1.24.3
pandas==2.0.1
Flask==3.0.0
elasticsearch==8.9.0
stanza==1.5.0
Werkzeug==3.0.0
sentence-transformers==2.2.2
Flask-Bootstrap==3.3.7.1
python-dotenv==1.0.0
spacy==3.7.1
flask-swagger-ui==4.11.1
passlib==1.7.4
bcrypt==4.1.2
PyJWT==2.8.0
Flask-JWT-Extended==4.6.0

